[add] 新增排班功能

+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -4,29 +4,36 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+/**
+ * 建立密碼重設表（password_resets）
+ */
 class CreatePasswordResetsTable extends Migration
 {
     /**
-     * Run the migrations.
-     *
      * @return void
      */
     public function up()
     {
+        if (Schema::hasTable('password_resets')) {
+            return;
+        }
+
         Schema::create('password_resets', function (Blueprint $table) {
-            $table->string('email')->index();
+            $table->string('account')->index()->comment('帳號');
             $table->string('token');
             $table->timestamp('created_at')->nullable();
         });
     }
 
     /**
-     * Reverse the migrations.
-     *
      * @return void
      */
     public function down()
     {
+        if (!Schema::hasTable('password_resets')) {
+            return;
+        }
+
         Schema::dropIfExists('password_resets');
     }
 }
